Remove leftover support for Python 2
Also update cmstestsuite.sh() description, as it no longer matched the
implementation.

+++ cms/locale/locale.py
@@ -223,9 +223,8 @@ class Translation:
             n /= self.PREFIX_FACTOR
             if n < self.PREFIX_FACTOR:
                 f = copy.copy(self.locale.decimal_formats[None])
-                # We need int because floor returns a float in py2.
                 # if 1000 <= n < 1024 d can be negative, cap to 0 decimals
-                d = max(0, int(2 - math.floor(math.log10(n))))
+                d = max(0, 2 - math.floor(math.log10(n)))
                 f.frac_prec = (d, d)
                 return (self.gettext(unit)
                         % babel.numbers.format_decimal(n, format=f,
